Turn off methane in dynlakes test
I was getting a gridcell-level methane balance check error when running
ERS_Lm25.1x1_smallvilleIA.IHistClm50BgcCropQianRsGs.cheyenne_gnu.clm-smallville_dynlakes_monthly.
I'm thinking it was due to ESCOMP/CTSM#43. So for now, I'm disabling
methane in this testmod. (I thought about changing it to an Sp case, but
(1) currently you're not allowed to run smallville in an Sp case, and
(2) it would be nice to exercise as much of the BGC code as possible in
this test, to exercise the dynamic landunit conservation code for BGC
variables.)