Calculate fire btran2 using updated h2osoi_vol
I had introduced a temporary, saved version of h2osoi_vol to achieve
bit-for-bit results in my testing. Now that I have verified that the
main changes are bit-for-bit, I'm changing this to use the actual
h2osoi_vol. This will change answers, because fire's btran2 will now
use the values of h2osoi_vol updated later in the driver loop.