Reinitialize fire btran2 to spval for all patches in each time step
This, or something like it, is important so that the fire code isn't
trying to use btran2 from earlier time steps for patches that were in
the exposed veg filter at one point but no longer are. The fix here is
not the cleanest way to fix this issue, but represents a minimal set of
changes to fix the issue. In a later commit, I'll try to fix this in a
better way (which should be bit-for-bit with this fix).

Resolves ESCOMP/CTSM#1153 (Fire btran2 is only computed for exposed veg
patches, but is used over all veg patches)