Limit fire btran2 to be <= 1
For the older version calculated in CNFireBaseMod, btran2 could
sometimes be roundoff-level greater than 1. Due to a conditional in
CNFireArea, these slightly-greater-than-1 values were being ignored when
computing btran_col, rather than averaging in a 1 value.

I haven't investigated the behavior in the new version in
CNFireLi2021Mod, but it seems like this could have the same issue. It's
also possible, though, that we'd exceed 1 by more than roundoff in that
version under some conditions, if h2osoi_vol > watsat in some
layers. The endrun call will tell us if that's happening. Note that I
plan to remove the endrun call after initial testing, because it seems
like the correct thing to do here is to limit the btran2 value to be no
more than 1 even if it exceeds 1 by more than roundoff.