Move calc_fire_root_wetness for CNFireLi2021 into the base type
The motivation here is to avoid future bugs: When the next fire version
comes along, we'll likely have it extend cnfire_base_type. But then we
would likely end up accidentally using the original
calc_fire_root_wetness rather than the 2021 version of that routine.

I don't love the decrease in locality that this refactor produces, but
it seems worth it to avoid that potential bug. And in any case, we'd
have to do something non-ideal when the next fire version comes along if
we want it to share the 2021 version of this routine. (We could have the
new version extend the 2021 version, but that leads to an even more
complex inheritance hierarchy than we already have for fire.)