Change CISM2%NOEVOLVE compsets to instead use SGLC
The use of CISM2%NOEVOLVE adds complexity and (perhaps most importantly)
increases the turnaround time of CTSM software testing. Long-term, we'd
like to use a data glc model for this purpose. For now, though, we're
going to use SGLC.

In addition to changing CISM2%NOEVOLVE to SGLC in compset definitions,
this also involves getting rid of the Gs designator (for SGLC) in
compset aliases, since now all I compsets are assumed to use SGLC,
unless they use a fully-active CISM (in which case they have a G at the
end of the compset alias).

Resolves ESCOMP/CTSM#1135