Move last CNNSTateUpdate1 use_fun matrix changes to CNPhenology
Remove 00fafae7a89429f2a0155b56762aa01a59dd95af changes in
CNNStateUpdate1 and put them in CNPhenology in LivewoodTurnover
with the other similar matrix updates. Note, that acc_ph HAS
to be TRUE in this case, as otherwise a Nitrogen balance error
results. With this the answers are the same as before.