Use python3 in more shebang lines
This is needed to run the python unit tests on cheyenne without
explicitly pointing to python3.

There are still more shebang instances throughout CTSM that I have not
yet changed, because I'm not positive if they should be changed
yet. (buildnml and buildlib should probably be changed in conjunction
with others through CESM, and I'm not sure about the usage of the
contrib tools.)