Remove "mec" from some glacier/ice variable names
It is misleading to have "mec" in variable names when the ice landunit
can actually have multiple columns *or* a single column, depending on
the glacier region behavior. This started as a rename of istice_mec to
istice, but then I realized that there were a bunch of other variables
that should also be renamed. However:

- I left "mec" in place for some things that truly were specific to
multiple elevation classes

- I left "mec" in place for surface datasets and the code that reads
them, partly because they are specific to multiple elevation classes,
and partly to avoid compatibility issues

- And I may have missed a few that should be changed

Resolves ESCOMP/CTSM#1294