Second round of moving parameters to param file.
hardcodep3 branch

Parameters moved:
pc
mu

The parameters "pc" and "mu" were defined in clm_varcon without the "_r8" appended and so
are not bfb when moved to the parameter file.
However, I reran the tests above for the baseline tag (PPE.n08_ctsm5.1.dev030) with the _r8
added for those parameters in clm_varcon and the new branch is bfb with that.  So I have some
confidence that the differences are just due to the missing _r8 for these two parameters.