Set o3 coefficients to 1 over noexposedvegp filter
This is needed for the sake of history diagnostics: Even though these
coefficients are only referenced within the exposed veg filter, the
corresponding history diagnostics are averaged over all veg patches. So,
to avoid having some "memory" of the last time a patch was
exposed (which is weird for diagnostics purposes and breaks exact
restart), we explicitly reset the coefficients to 1 for all points in
the noexposedveg filter.