Added boolean namelist parameter generate_crop_gdds.
In group cnphenology. Default false. Set to true (along with use_cropcal_streams) in order to override crop harvesting logic and to instead harvest the day before the next sowing date. Used to generate growing-degree day outputs that can be used with an external script to generate new GDD requirement ("cultivar") files.

I am not allowing generate_crop_gdds set to true if use_cropcal_streams is false. There's no theoretical reason this should be disallowed (would harvest the day before next sowing window began), but I don't see a point to it, and it's extra code to write so I'm skipping it.