Removed cropplant_patch, a vestige of early planning re: crop rotation.
In the absence of a system for crop rotation, cropplant was only used to prevent a crop from being re-planted in a given year after it's already been harvested that year (with the beginning of the "year" differing between hemispheres). This is redundant with my new variable crop_inst%growingseason_count, which tracks the number of growing seasons that have begun in a given calendar year (i.e., since 1 Jan 00:00).

See discussion at #1500 (https://github.com/ESCOMP/CTSM/issues/1500).