get_curr_calday: only apply the hack for day 366 if requested
This hack is needed when the result will be used for a call to
shr_orb_decl, but should not be used in other situations. (There were
some calls where the result was not directly used in a call to
shr_orb_decl, but it looked like the result should stay consistent with
other variables that are used directly in a call to shr_orb_decl.)

This also adds some unit tests. The two unit tests that test Dec 31 on a
leap year failed before making the changes to clm_time_manager in this
commit.