Add explicit function get_prev_days_per_year
And rename get_days_per_year to get_curr_days_per_year to be explicit.

This fixes the previous commit, where the wrong type was used for the
offset argument, in addition to making the use more explicit.

The behavior for AnnualFluxDribbler is as in the last commit, with a
change from master. (The message from the last commit was the following,
which remains effectively the same here, though with implementation
details changed: On the last time step of the year, get_days_per_year
would use the "current" date (i.e., time 0 of the next year) and so
would give the number of days per year in the next year. But what we
actually want is the number of days per year in the just-ending year; we
achieve this by using an offset of -dtime in the call to
get_days_per_year.)

For other uses, I have not yet determined what is correct, but I am
maintaining the previous behavior.