Added annual outputs of sowing and harvest dates.
SDATES and HDATES, respectively. Sowing dates are on new dimension mxgrowseas (maximum number of growing seasons begun this year; currently hard-coded to 1). Harvest dates are on new dimension mxharvests, which is mxgrowseas+1. The lengths of these dimensions are public constants of clm_varpar.

Cherry-picked from various commits on other branches in my fork.