Change CropPhenology to look at time as of the START of the time step
Previously, CropPhenology looked at time as of the END of the time step.
This was somewhat problematic, particularly because it meant that the
last time step of the year was considered Jan 1, and so crops with a
planting window beginning Jan 1 could be planted at the end of the
previous year rather than the start of the new year. This was becoming
particularly problematic in the context of Sam Rabin's upcoming
prescribed sowing date work (see
https://github.com/ESCOMP/CTSM/issues/1623 and
https://github.com/ESCOMP/CTSM/pull/1616 for some discussion).

I thought about whether the calls to `get_curr_date` in
`CNPhenologyClimate` and `CropIncrementYear` should also be changed to
`get_prev_date`. These are used for the 20-year running average GDD
accumulation (`CropIncrementYear` is used for updating
`nyrs_crop_active_patch`, but that in turn is just used for the GDD
accumulation). But these really do feel like they are appropriate to
happen at the end of the year – not the first time step of the new
year. (One reason why the current end-of-year timing is better is for
the sake of diagnostics – so that the 20-year averages have been updated
before writing an end-of-year history file.) It feels like the ideal
time for it to happen would be *after* the other crop phenology stuff,
but I don't think it's a big deal for it to happen before the other crop
phenology stuff in the last time step of the year, as is currently the
case.

Resolves ESCOMP/CTSM#1623