Tweak S f09 PE layout
With this small node count, having datm on its own node only makes
things a little faster, and adds significantly to the cost (and might
lead to slightly longer queue wait times), so I don't think it's worth
it. However, datm is slightly faster with just one node rather than 4,
so I'm maintaining that aspect of the PE layout.