Use smaller PE count for PFS test
I compared the detailed timing files from this new
PFS_Ld10_PS.f19_g17.I2000Clm50BgcCrop.cheyenne_intel test with the
original PFS_Ld20.f09_g17.I2000Clm50BgcCrop.cheyenne_intel test. For the
most part, the relative timings of different pieces are about the same
in both - at least for the major culprits, which is what I compared. So
this is good: I feel like we can use the new test to still give
indications of when the timing of a particular section changes
significantly, and be reasonably confident that the same result would
hold for an f09 test with higher processor count.