For CLM45 apply peaklai to aleaf in grainfill
This makes the crop allocation code in
NutrientCompetitionCLM45defaultMod consistent with that in
NutrientCompetitionFlexibleCNMod. This will be helpful so that I can
extract these into a shared subroutine.

From a 5-year run of IHistClm45BgcCrop at f19_g17 resolution, it seems
like this won't change answers: whenever peaklai is 1 during the
grainfill period, aleaf is already exactly 1e-5. I don't understand why
this is always the case; maybe, in Clm45, if peaklai isn't reached
before grainfill then it is never reached that growing season? I have
come to suspect that the difference between Clm45 and Clm50 in this
respect may actually be due to whether you're running with FUN, since
running with FUN changes some relevant driver sequencing. So it's
possible that this would change answers for the atypical situation where
someone was running with FUN but not FlexibleCN; in this case, I think
this change would be the right thing to do.