Extract the calculation of crop allocation fractions into its own subr
This way we can share this code between the different nutrient
competition methods rather than having it duplicated. It also gives us a
path towards skipping these calculations when running with AgSys.

One change I made in doing this refactor was changing allocation
fractions to always sum to 1, even in pre-emergence and when not
croplive... this shouldn't make any difference outside of the allocation
fractions themselves (and note that those aren't output to history
files, so I don't think this should lead to any answer changes).