Revert "Allow time-varying C:N ratios"
This reverts commit 51241e8c3e5236ba0c0cb6adde78398ae4cd8dd2.

After further discussion, it isn't clear if we really want to get
time-varying C:N ratios from AgSys, and if we do, it isn't clear what
those should be and in which places we should use the time-varying vs.
time-constant C:N ratios. So for now we're sticking with the use of
pft-constant C:N ratios; we may revisit this later.