Change default bld dir from /bld to /tool_bld for sys test to work
The long version...
System test
MKSURFDATAESMF_P144x1.f10_f10_mg37.I1850Clm50BgcCrop.cheyenne_intel.clm-default
relies on the ability to build the mksurfdata executable anywhere. I had
named this directory (that can go anywhere) /tool_bld to distinguish
from /bld that appears during the sys test in the same case directory in
prep for submitting the ctsm run. Due to changes coming in the next
commit that make the sys test generate and submit the jobscript that
runs the mksurfdata executable (instead of spelling out the command
inside the sys test script), I had to make the name of the default bld
directory the same as the one expected by the sys test.