Introduce lilac_init_pio_mod
This allows us to have a portion of shr_pio_mod that is truly shared
between the drivers and a portion that is driver-specific.

This file is currently identical to
components/cpl7/driver/main/init_pio_mod.F90, except (1) the module name
has been changed to lilac_init_pio_mod (rather than init_pio_mod), (2)
all init_pio_* routines have been renamed to lilac_init_pio_*, and (3)
this description comment has been added.

This needs to be coordinated with a branch in the CESM_share repository.