Apply dust branch based off of release-clm5.0.34 to dev branch
Do a diff of the dmleung DUSTEMIS branch based off of release-clm5.0.34
to ctsm5.1.dev106 and apply the diff after updating the paths for
namelist_definition and lnd_import_export.F90 in the diff.

git apply -3 dust.diff

Diffs with conflicts:
	modified:   src/biogeochem/DUSTMod.F90
	modified:   src/biogeophys/SoilStateInitTimeConstMod.F90
	modified:   src/main/controlMod.F90
	modified:   src/cpl/mct/lnd_import_export.F90

The changes in lnd_import_export were to change a check for negative
LW that seems to be removed in the latest version.