Don't overwrite soil properties just because pctlnd_pft is small
It doesn't seem like there's any reason why we need to overwrite soil
properties with a default value just because pctlnd_pft is small. Doing
so was fine when we turned those grid cells into wetland, but now that
we aren't necessarily doing that, it makes sense to keep valid soil data
wherever we have it.

This could change answers in places where we have valid soil data
despite pctlnd_pft being near 0.