Don't reapply user-mod in run_neon clone, and don't set DATM_YR_END
The base case already has the NEON user-mods applied in it, so
we don't need to reapply in the cloned case.

Don't set DATM_YR_END, let it be set in the default user-mods
and the specific user-mods for the site. The only advantage of
setting DATM_YR_END in run_neon is if your run-length is shorter
than what's in the user-mods, and it only saves downloading
the extra files and a little bit of time to read them at
initialization. So there's no need to have logic to handle that
here.