Remove / rework some redundant adjustments based on pctlnd_pft
Do some rework of adjustments now that we are mapping pctnatveg and
pctcrop with DSTAREA rather than FRACAREA normalization - so they end up
as percent of the grid cell rather than percent of the land area. The
end result won't be identical in all cases, but in practice it should do
the same thing as before.

- Zero out pctnatveg and pctcrop if their areas are less than 1e-6,
rather than the previous logic of zeroing these if pctlnd is less than
1e-6. (This zeroing is now done in mkpft rather than in the main
mksurfdata program.)

- In deciding whether to set pct_nat_pft or pct_cft to a fixed value,
only check for tiny values of pctnatveg / pctcrop, rather than checking
for a tiny value of pctlnd (it should now be the case that a tiny value
of pctlnd implies tiny values of pctnatveg and pctcrop as well).

- Remove the check for no vegetation outside the pft mask. This check no
longer seems important and risks doing more harm than good if the
threshold for this check differs slightly from the threshold used to
zero pctnatveg / pctcrop.