Repl code that sort-of worked: transient lake+urb use actual first yr
With code that I had added before in gen_mksurfdata_namelist.py,
mksurfdata_esmf aborted when trying to continue from 2015 to 2016.
To troubleshoot, I tried mksurfdata_esmf from 2016 to 2100 and found
that urban was not starting with the correct 2016 file.
I modified the code for both urban and lakes to use the method already
in use for transient pfts. I tested 1850-2100 and 2016-2100 and they
worked.