Loosen tolerance on near-zero truncation of h2osoi_ice and h2osoi_liq
In UpdateState_TopLayerFluxes, there is a somewhat arbitrary epsilon
used to determine if h2osoi_ice and h2osoi_liq are close enough to zero
that they should be truncated to zero. If the state remains negative
after this truncation, we deem this to be a problem. It seems that this
tolerance is occasionally exceeded, leading runs to abort. Since this
tolerance is somewhat arbitrary, we will loosen it by an order of
magnitude.

Resolves ESCOMP/CTSM#1979

Note that, although the issue was only reported for h2osoi_ice, I am
changing the tolerance for both liq and ice so they remain consistent.