Put do_harvest back where force_harvest was removed from cphase_leafemerge conditional.
This represents a bug fix over the previous behavior: There were situations where force_harvest would be false with do_harvest true, meaning that the latter was ignored. This fix changes the timing of crop phase transitions, but rarely, and not by much. Does not seem to affect GDD-generating runs.