Added evenly_split_cropland option to fsurdat_modifier.py.
This could also be accomplished by using dom_pft to specify an equal value for all crop PFTs. However, this method is more robust to user error (we don't have to rely on the user correctly writing out a long string of numbers) and code updates (if more crop PFTs are added).