Make cmds_to_setup_conda() more robust.
For tests that invoke cmds_to_setup_conda(), manually calling the script invoking that function (e.g., case.build for FSURDATMODIFYCTSM) could fail if doing so with a conda environment already activated. The problem is that
    conda run -n ctsm_pylib
seems to not actually use ctsm_pylib if, for instance the conda base environment is active. Instead doing
    CONDA_PREFIX=
    conda run -n ctsm_pylib
seems to work.