All accumulator types can now be manually reset.
Previously it was just runaccum, which would reset a point-level if that point-level had the value -99999._r8. That's potentially unsafe, since theoretically it could be a valid value. This behavior has now been replaced by looking for a .true. value for the point-level in the new "reset" array.

Accumulator types timeavg and runmean can also look at "reset" for a manual reset. These types were not able to be manually reset before. Note that these have some different reset behavior compared to runaccum: Whereas runaccum can't reset and update in the same call of its update_accum_field() subroutine, timeavg and runmean can (and do).

runmean's behavior is also a bit different from the others in that it will "remember" reset requests that occurred when a point was inactive. The others already had reset blocks that were embedded in a check of "active," so that behavior is preserved for them.

All existing unit tests pass. No new unit tests have been added yet.