Actually don't use TemperatureType%flush_gdd20 at all.
Having that alongside flush_gdd20 from clm_varctl means that they can disagree when restarting, and it's not obvious which should win.

Note that the user being able to request a flush at all is dangerous. If they start a run with it true, they might continue that run without setting it to false. Instead, the restart decision should be made by the model---if a gridcell has different prescribed gdd20 season from what its restart file says, then that gridcell should be flushed.