Keep GLACIER_REGION=1 unused
We realized that the renumbering of regions could cause silent wrong
behavior if a user was using an old user_nl_clm file that specified
behaviors for the old 4 regions. So instead, we're keeping the old
numbering for the Greenland and Antarctica regions and keeping
GLACIER_REGION=1 unused for now.