Changes needed to get pFUnit tests building and running on my Mac
I don't understand how the previous version has been working on derecho:
it seemed to be relying on things set in the share subdirectory, but I
was finding that the settings there didn't carry back to this top-level
location. Maybe the derecho build relies on some environment variables
or compiler flags that are set by modules??? In any case, I had a lot of
problems getting it to work on my Mac. This version works on my Mac.
It's possible that not all of these changes are needed, but this new
version makes more sense to me.

A lot of these changes were copied from the top-level CMakeLists.txt in
the share directory.