Remove inclusion of PIO in the unit test build
This came from working with Adrianna to get things working on her Mac.
She used a version of ESMF with internal PIO, so I don't think we can do
this find of PIO. And it turns out that it works to remove this, both in
her build and in mine, as long as we remove the unneeded shr_pio_mod.F90
from the unit test build.

I don't understand why this works: I thought it would need `-lpio` or
something like that in the link line since ESMF depends on the PIO
library. But maybe pio is folded into the esmf library or picked up
dynamically through rpath or something like that???

The current state - with PIO not included in the unit test build - means
that we may not be able to use PIO stuff directly from unit tests, but
as Adrianna points out, that's probably a good thing anyway. I'm also
concerned that it might be fragile - so we might need to add it back in
later. But since it isn't needed for now, I'm not even sure that what's
there works.