Allow timemgr_reset to be called safely if the timemgr was uninitialized
Without this change, the unit tests of clm_time_manager were aborting
due to trying to destroy an ESMF clock that hadn't been initialized.
Other pieces of this reset routine don't appear to be causing problems,
but it seems most robust to just skip the whole thing if the time
manager hasn't been initialized.