Fix time-related unit tests to work with real ESMF library
There is a behavior difference between the real ESMF library and the
ESMF WRF Timemgr we had been using previously, in how prevTime is
defined: the ESMF WRF Timemgr defined prevTime using a -dtime offset
from currTime; the real ESMF library instead defines prevTime by
tracking the previous value of currTime. Therefore, now that we are
using the real ESMF library, if a unit test relies on prevTime (as is
the case, for example, with test_year_position_start_of_timestep in
test_dynTimeInfo.pf), it is now important that any setting of the time
be done in a way such that there is a call to ESMF_ClockAdvance, since
that is the only way that prevTime will get set correctly. Thus, this
commit updates the implementation of for_test_set_curr_date so that it
involves a call to ESMF_ClockAdvance.

Making this change in the implementation of for_test_set_curr_date also
required some other changes in some unit test code:
- I needed to remove the call to unittest_timemgr_set_curr_date at the
  end of unittest_timemgr_setup. This call now leads to an increment of
  the time step count, which was causing failures in some time manager
  tests that check the time step count. And it turns out that (at least
  with the other changes here), the call to
  unittest_timemgr_set_curr_date was unnecessary: it wasn't actually
  leading to any change in time from what was set in the initialization
  of the time manager.
- I needed to change some set_date calls to use a year of 2 instead of
  1. This is needed for calls setting the date to January 1 and the time
  to 0: It is no longer allowed to set the date to yr,mon,day,tod =
  1,1,1,0, because the new method sets the date/time to one time step
  earlier and then advances the clock, and it is illegal to set the
  date/time to one time step before 1,1,1,0.