Make snow resetting dependent on glacier region
For now this is a hack, hard-coded to only do snow resetting over
Greenland. If we want this behavior long-term, we should introduce a new
glacier region behavior namelist flag controlling this behavior; this
would require changes in glcBehaviorMod.F90 in addition to other changes
to add the new namelist flag. (The changes in this commit outside of
glcBehaviorMod.F90 could remain as they currently are.)