Check if sharedlib_only rather than for existence of subset_data outputs.
build_phase gets called twice:
- once with sharedlib_only = True and
- once with model_only = True
Because we only need subset_data run once, we only do it for the sharedlib_only call. We had originally checked for the existence of the subset_data outputs, but that might lead to a situation where the user expects subset_data to be called but it's not. Better to run unnecessarily (e.g., if you fixed some FORTRAN code and just need to rebuild).